+++ src/main/java/racingCar/domain/Winners.java
@@ -14,29 +14,6 @@ public class Winners {
         return this.cars;
     }
 
-    public static Winners getWinners(List<Car> carList) {
-        return getWinners(carList, getMaxDistance(carList));
-    }
-
-    private static Winners getWinners(List<Car> carList, Distance maxDistance) {
-        List<Car> winners = new ArrayList<>();
-
-        for(Car car : carList) {
-            if(car.isMax(maxDistance)) {
-                winners.add(car);
-            }
-        }
-        return new Winners(winners);
-    }
-
-    private static Distance getMaxDistance(List<Car> carList) {
-        Distance max = new Distance();
-        for(Car car : carList) {
-            max = car.max(max);
-        }
-        return max;
-    }
-
     public List<String> getNames() {
         List<String> names = new ArrayList<>();
         for(Car car : cars) {
