+++ src/test/java/racingCar/CarsTest.java
@@ -2,10 +2,12 @@ package racingCar;
 
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.Test;
-import racingCar.domain.Car;
-import racingCar.domain.Cars;
+import racingCar.domain.*;
 
+import java.util.Arrays;
 import java.util.List;
+
+import static org.assertj.core.api.Assertions.assertThat;
 
 public class CarsTest {
     @Test
@@ -24,5 +26,18 @@ public class CarsTest {
         Cars cars = new Cars(list);
 
         Assertions.assertThat(cars.value()).hasSize(3);
+    }
+
+    @Test
+    void 우승자_출력() {
+        Car pobi = new Car(new Name("pobi"), new Distance(5));
+        Car crong = new Car(new Name("crong"), new Distance(2));
+        Car honux = new Car(new Name("honux"), new Distance(2));
+
+        Cars cars = new Cars(Arrays.asList(pobi, crong, honux));
+
+        Winners winners = cars.getWinners();
+
+        assertThat(winners.value()).contains(pobi);
     }
 }
